Im pushing 2 modules
The first modil is Piece.
Here is the creation of the pieces, the shape defines, moving piece.
Here is anotwher thing - Predict.
It is the prediction if the falling pieces.
Here the moving if the Prediction is done.

The second modul is Shape.
Here the shape defines, moving the construction, dropping.